function fix

+++ build-rpi4.sh
@@ -75,8 +75,8 @@ checkfor_ubuntu_image () {
         fi
     fi
     # Symlink existing image
-    [ ! -f /build/source/${ubuntu_image} ]; then 
-    ln -s /$ubuntu_image /build/source/
+    if [ ! -f /build/source/${ubuntu_image} ]; then 
+        ln -s /$ubuntu_image /build/source/
     fi
 }
